Add nested JSON fields to plain table and create core util for nested JSON fields
