+++ packages/frontend-core/src/utils/schema.js
@@ -1,5 +1,6 @@
 import { helpers } from "@budibase/shared-core"
 import { TypeIconMap } from "../constants"
+import { convertJSONSchemaToTableSchema } from "./json"
 
 export const getColumnIcon = column => {
   // For some reason we have remix icons saved under this property sometimes,
@@ -23,4 +24,26 @@ export const getColumnIcon = column => {
       : TypeIconMap[type]
 
   return result || "Text"
+}
+
+export const addNestedJSONSchemaFields = schema => {
+  if (!schema) {
+    return schema
+  }
+  let jsonAdditions = {}
+  Object.keys(schema).forEach(fieldKey => {
+    const fieldSchema = schema[fieldKey]
+    if (fieldSchema?.type === "json") {
+      const jsonSchema = convertJSONSchemaToTableSchema(fieldSchema, {
+        squashObjects: true,
+      })
+      Object.keys(jsonSchema).forEach(jsonKey => {
+        jsonAdditions[`${fieldKey}.${jsonKey}`] = {
+          type: jsonSchema[jsonKey].type,
+          nestedJSON: true,
+        }
+      })
+    }
+  })
+  return { ...schema, ...jsonAdditions }
 }
